Trim the output line, since we're erasing to EOL anyway
Currently renderTemplate ALWAYS produces a string of the length you specify,
even if it has to pad stuff to produce that.  This renders the
"erase-to-eol" command meaningless, so we either should ditch that or we
should trim the renderTemplate output. Right now I'm opting for the
latter, but testing the former would be useful too.